+++ controller/index.go
@@ -1,21 +1,6 @@
 package controller
 
-import (
-	"net/http"
-
-	"github.com/unrolled/render"
-)
-
-//Render :render data var
-var Render *render.Render
-
-// Init :init controller methods
-func Init() {
-	Render = render.New(render.Options{
-		Directory: "templates",
-		Layout:    "layout",
-	})
-}
+import "net/http"
 
 //HandlHome :handle "/"
 func HandlHome(w http.ResponseWriter, req *http.Request) {
